Moved useLayout/useSidePanel to utils
